+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.9.1
 
 ### Added
-
+- `GetSolid()` method on GeometricElement that returns the Csg solid.
 - `Polygon.ToTransform()`
 - `Elements.Anaysis.AnalysisImage`
 
@@ -13,6 +13,7 @@
 
 - `AnalysisMesh` now handles single valued analysis.
 - `Polygon.Split()` can now handle polygons that are not in the XY plane.
+- Leave the discriminator property during deserialization.  It will go to AdditionalProperties.
 
 ### Fixed